% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ railroadR }
\alias{ railroadR }
\alias{ railroadR }
\title{ Railroad Diagrams for R }
\description{
Generate railroad diagrams in R with the convenience and infrastructure of htmlwidgets, or create static SVG railroad diagrams for use in non-htmlwidget situations.
}
