% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/railroad.R
\name{railroad-shiny}
\alias{railroad-shiny}
\alias{railroadOutput}
\alias{renderRailroad}
\title{Shiny bindings for railroad}
\usage{
railroadOutput(outputId, width = "100\%", height = "400px")

renderRailroad(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width,height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a railroad}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
  is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using railroad within Shiny
applications and interactive Rmd documents.
}

