% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/railroad.R
\name{railroad}
\alias{railroad}
\title{Railroad Diagram}
\usage{
railroad(..., width = NULL, height = NULL)
}
\arguments{
\item{...}{string arguments to form the diagram spec}

\item{width,height}{width and height in any valid CSS unit for
               the htmlwidget container}
}
\description{
Create interactive railroad diagrams for use in html.
}
\examples{
library(railroadR)

# examples from railroad-diagrams
#   https://github.com/tabatkins/railroad-diagrams/blob/master/example.html

railroad(
  NonTerminal('IDENT'),
  '('
)

railroad('@', NonTerminal('IDENT') )

railroad('#', NonTerminal('IDENT') )

railroad(
  Choice(0, 'u', 'U'),
	Choice(0, 'r', 'R'),
	Choice(0, 'l', 'L'),
	'(',
	Choice(1,
		Optional(NonTerminal('WS')),
		Sequence(
			Optional(NonTerminal('WS')),
			NonTerminal('STRING', '#string'),
			Optional(NonTerminal('WS'))),
		Sequence(
			Optional(NonTerminal('WS')),
			OneOrMore(
				Choice(0,
					NonTerminal('not " \\' ( ) WS or NPC'),
          NonTerminal('escape'))),
          Optional(NonTerminal('WS'))
		    )
		),
  ')'
)

railroad(
" Choice(1, '+', Skip(), '-'),
  Choice(0,
         Sequence(
           OneOrMore(NonTerminal('digit')),
           '.',
           OneOrMore(NonTerminal('digit'))),
         OneOrMore(NonTerminal('digit')),
         Sequence(
           '.',
           OneOrMore(NonTerminal('digit')))),
  Choice(0,
         Skip(),
         Sequence(
           Choice(0, 'e', 'E'),
           Choice(1, '+', Skip(), '-'),
           OneOrMore(NonTerminal('digit'))))
"
)
}

